requirements.txt
tensorflow>=2.4.0
numpy>=1.19.2
matplotlib>=3.3.2
pillow>=8.0.0
opencv-python>=4.5.1
face-recognition>=1.3.0 